Fix hashes on big-endian machines.
In the standard padding mechanism for hashes, we append the length as a number
with twice as many bits as the quantum used.  We have to treat this number as a
pair of quantum-sized numbers.  Because of this, the first value (that is, with
index 0) is always first in memory.  On big-endian machines, this is the least
significant value and therefore should instead be second.  This commit fixes it
so that the value is correct on both big-endian and little-endian machines.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.ath.cx>