Stop doing odd pointer tricks.
When libdrew needs to get a plugin function from dlsym, it has to cast from a
void * to a function pointer.  This is not strictly allowed by ISO C or POSIX,
but it is functionality required by XSI.  We used to do odd things with void *
to try to avoid the warning, but that led to strict-aliasing warnings, which can
actually break things.  We now just simply cast and assume it will work.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.ath.cx>