Make hash maintenance test like other hash tests.
In general, the hash tests return a boolean true if successful, false otherwise.
The hash maintenance tests originally used an int as a result which was 0 if
successful, positive otherwise.  As a result, it was easy to misuse the
maintenance tests, so convert them to have a boolean return value.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>