rijndael: remove useless "inline".
GCC complains that our inline pure virtual functions are never defined (since
they're pure virtual), so remove the "inline" to shut it up.  This is not likely
to have a significant impact on performance; even if it does, these
implementations are only used for decryption and only for the rijndael plugin,
not the aes one.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>