arc4stir: improve performance.
This version of the arc4stir PRNG has significantly increased performance (from
0.790 MiB/s to 90.473 MiB/s).  The algorithm only stirs approximately 2**20
bytes instead of 2**8, but this is still less than the OpenBSD arc4random
algorithm.  It also now drops approximately 3072 bytes after adding random data,
which ideally would not be necessary, but is again more conservative than the
arc4random algorithm.  It also tries to acquire some random data when stirring,
independent of what is provided with the seed interface.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>