libdrew: fix crash.
drew_loader__is_valid_id used to check for the entry having a non-NULL name
attribute.  However, that should never occur, since drew_loader_load_plugin
always strdups a non-NULL name.  Anyway, the name attribute is the plugin name,
and this is never exposed outside of the loader itself.  What is likely to
occur, and what is important to check for, is that the *algorithm* name (aname
attribute) is non-NULL; the check has been changed to this.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>