util: handle byte-swapped data better.
Previously, wrong-endian data was copied to the destination buffer and then
reversed.  This worked fine if the data length was a multiple of the chunk size,
but if not, some data would be lost.  The new strategy is not to memcpy first,
but instead to manually copy the data across.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>