test/framework: don't lose error condition.
Previously, we just shifted the result left by one bit.  However, if we're
running a large number of tests, we can lose the fact that there's an error if
the error is early on.  This commit introduces a way to ensure that doesn't
happen at the expense of making the result code slightly less useful.  However,
since we don't use the result code to generate the list of errors, this is
largely academic.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>