sha512: remove large tests.
NIST has published several testcases that involve many iterations that take a
long time.  While this is not a problem on some systems (3 minutes on a 2.0GHz
Core 2 Duo), on older systems this can be prohibitive (3 hours 15 minutes on a
333MHz Ultra 5).  Since these tests have been moved to the external testing
routines (and were in fact the reason for creating the external testing
routines), this does not cause any risk of incorrect code.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>