rc4: batching and performance improvements.
This commit introduces a keystream buffer that holds the 256 bytes that are now
generated at a time.  As a consequence, the algorithm's i variable can be
subsumed into the loop index.  Because the loop index is guaranteed never to
exceed 255, the need for converting it into a uint8_t is now gone as well.  The
speed difference is negligible for a uint8_t-templated keystream, but causes a
significant speed increase (188 MiB/s to 216 MiB/s) for an int-templated
keystream, which is now the default.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>