libdrew: return id on plugin-loading success.
In the old implementation, we would return a special ID value that signified not
a plugin, but a library.  Because we don't keep plugins and libraries in the
same structure anymore, instead we return the ID to the first plugin, which can
be used in drew_loader_get_nplugins, which is the only place such a library ID
is valid anyway.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>