include/util: use htonl and friends for Convert.
The Convert functions for BigEndian conversions now use htonl and friends for
efficiency reasons.  Glibc provides assembly versions of these in some cases,
and where it does not, it provides code that GCC can optimize into a single
instruction if that is available.  Presumably other systems provide similarly
optimized code.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>