util: consolidate endianness code.
This commit converts the BigEndian and LittleEndian classes into little more
than templates of an Endian class.  This means that code will no longer be
duplicated, leading to difficult-to-find errors.  They are not simply typedefs
because there are valid optimizations that work on some endianness and not
others, such as the htonl family of functions, that we want to continue to use.
There are no noticable performance consequences.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>