tools: rework compare-speed.
This commit introduces several improvements.  It does not error out when one
side has an algorithm and the other does not.  It does not matter if items are
in a different order, as long as different implementations of the algorithm
retain their relative order.  It also does not require that the files be exactly
the same length in lines.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>