Split out external test code into its own file.
The external test code is great for algorithms that are designed to be
repeatable, but this is not the case for PRNGs.  Since PRNGs will need their own
external test code, split the standard external test code out so it can be used
only by those test binaries that need it.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>