Add a bignum implementation using OpenSSL.
This is more of a test case to see what kind of operations can be expected out
of a bignum library than an actual useful implementation.  Also, since we'll be
emulating a lot of the OpenSSL implementation, it is rather silly to actually
depend on OpenSSL.  Nevertheless, the implementation does work and the public
key tests do pass with it.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>