libdrew: handle plugins with no implementations.
If the first plugin loaded contains no implementations, the reallocation of the
plugin information array may return NULL (since 0 bytes were requested to be
allocated).  In this case, the -ENOMEM error will be returned, which isn't
really what we want.  It also differs from the behavior if a plugin other than
the first contains no implementations.  For simplicity's sake, simply allocate
one more entry than necessary so that we always get a non-NULL pointer.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>