x509: handle v1 certs with short version numbers.
The previous logic only handled version 1 certificates with an omitted version
number if the serial number was too large to be held in a ssize_t.  However,
there are certificates (GTE CyberTrust, I'm looking at you) which have a small
serial number for their v1 certificate.  Previously, we'd get an very large (>
100) value for the certificate version number; now we get the proper value.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>