openpgp: implement a keystore.
The idea behind the keystore is to store data in a format which preserves all
the implementation details in an easy-to-parse format.  Since other programs
will not use this format, the data should be able to be relied upon.  This
avoids having to check signatures more than once, etc.  At some point, a
Berkeley DB backend will be implemented which may allow delayed loading.

This also allows one central location to query the keyring for whatever data
might be available.  Searches based on user IDs and such are forthcoming.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>