hmac: minimize allocations.
The new memory allocation routines have a non-zero overhead compared to the
standard library functions because of the additional bookkeeping that needs to
be done.  This is not noticable in most cases, but hmac does a large number of
allocations and as a consequence, users of hmac, such as sp800-90, perform much
more slowly (88% in some cases).

This patch makes buffers a fixed size, which should be large enough for all
currently-implemented algorithms, which actually improves performance over the
pre-drew_mem implementation.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>