Use hidden symbols.
Hiding symbols with a linker script is not supported except on Solaris and with
the GNU linker.  Therefore, for many platforms, private symbols (such as all C++
symbols) are exported.  This commit introduces two pairs of macros for exporting
and hiding symbols and uses them throughout the code.  The benefit here is that
we can also hide module-specific info functions which would otherwise have been
exported.  Also, this functionality will be needed for an efficient
implementation of the OpenPGP library.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>