Add structure for storing subpackets.
We store subpackets in two ways: as a byte stream and as a set of parsed
subpackets.  Previously, each structure holding subpackets had to define the
four fields to store these two representations.  This commit unifies the four
fields into a structure and each containing structure includes two copies of it:
one for hashed subpackets and one for unhashed.  This also allows one simple
copy implementation to be used for copying all subpacket structures.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>