Fix definition of macro.
Since each set of IDs gets its own chunks, we want to round up to the nearest
even number, but we then need to divide that by two so we get the number of
chunks, not the number of ID slots.  This also has the side effect of decreasing
the size of the keystore.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>