Add an AESNI implementation.
This plugin uses the AES instructions on newer Intel processors.  The plugin
will build successfully regardless of the type of computer and regardless of the
presence or absence of the AES instructions at compile or runtime.  If the -maes
argument is not provided to the compiler, the implementing code will not be
compiled in; if the code is compiled in and the AES instructions are not
detected at runtime, the plugin will be empty, just as if the code had not been
compiled in.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>