Fix querying a zero number of implementations.
A plugin can contain zero implementations; this is common when the hardware does
not support the functionality required of the plugin.  We should return zero
when querying the number of implementations in such a plugin.  Previously, we
did not allocate an ID for such dummy plugins, so querying the number of
implementations would return an incorrect value.  Now, we allocate an ID that is
useless for anything else but can be queried for the number of plugins.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>