serpent: implement arbitrarily-sized keys.
This commit implements support for any key that is a multiple of 8 bits.
Previously, the implementation only supported keys of a length that was a
multiple of 32 bits.  Intentionally included in this commit is a set of
Sosemanuk test vectors that use a 40-bit key.  Since Sosemanuk uses the Serpent
key scheduling algorithm (among other components), this verifies that the
implementation does, indeed, work.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>