openpgp: properly check key ID when synchronizing.
Due to a thinko when converting the key code to C++, we were trying to validate
every user ID signature as a self-signature.  Unfortunately, this led to trying
to use an RSA key as a DSA key, which caused libtommath (the default bignum
library) to get caught in a nasty loop.  This commit fixes the problem so the
comparison is correct.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>