x509: compute digest of certificate.
This digest is the same digest that the digital signature uses.  It is not the
same as the fingerprint, which is simply the digest of the entire DER-encoded
certificate.  The X.509 code must generate the digest so the software does not
need to parse the certificate twice, whereas with the fingerprint, no special
parsing is needed.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>