Create self-signatures on demand.
Most signatures are not self-signatures.  We were previously using 88 bytes (on
amd64) of memory in each signature that was not used in the majority of cases.
This leads to a very large amount of wasted memory when processing large
keyrings such as the Debian keyring.  This commit uses a pointer and creates the
self-signatures on demand instead.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>