drew-openpgp: write the keystore to disk once only.
With the file backend, if we write a keystore more than once, we get the entire
keystore concatenated with itself multiple times.  This is problematic because
the header chunk is not allowed in the middle of a file, since, unlike other
key chunks, it has no subsequent chunks and therefore the data will be
misinterpreted and an oversize allocation will occur.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>