mem: allow adjusting secure memory flags.
Most cryptographic libraries do not offer secure memory allocation.  Drew does,
but the default maximum that can be locked into memory on most GNU/Linux systems
is 64 kB.  This is a problem for some algorithms (e.g. GCM) that use large
tables in secure memory, since the total amount to be locked into memory exceeds
that value.  This commit introduces an API to modify the behavior of the secure
memory allocator either to not lock memory or to succeed if locking memory
fails, which allows users to adjust this behavior.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>