block: allow passing object for testsuites.
Often the constructor of a block cipher will take a parameter, such as the
number of rounds.  The standard BlockTestCase class did not allow that, and as
such the test had to be done by hand.  This commit allows passing a pointer to a
block cipher to the testcase so the algorithm can be successfully tested.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>