Don't #define __BSD_VISIBLE on FreeBSD.
Because util.h is not included as the first header in any given file, we have
probably already included <sys/cdefs.h>, setting up all the appropriate macros.
If we then define __BSD_VISIBLE, we are going to start including other files
which depend on the types that that #define enables, except that since we hadn't
set it before, those types don't exist.  This breaks at least one of the prng
files.  Furthermore, FreeBSD doesn't seem to need it.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>